Move userGetter to commonPromises
